Fixing bugs and extending functionality

+++ src/main/java/com/_agents/library/controller/AuthorController.java
@@ -1,17 +1,20 @@
 package com._agents.library.controller;
 
-import com._agents.library.exceptions.AuthorNotFoundException;
-import com._agents.library.entity.Author;
-import com._agents.library.repository.AuthorRepository;
+import java.util.List;
+import java.util.Optional;
+
+import com._agents.library.exception.UniqueIdentifierModificationException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import com._agents.library.entity.Author;
+import com._agents.library.exception.AuthorNotFoundException;
+import com._agents.library.repository.AuthorRepository;
 
 @RestController
-@RequestMapping("/library")
 public class AuthorController {
 
-    private final AuthorRepository repository;
+    private AuthorRepository repository;
 
     AuthorController(AuthorRepository repository){
         this.repository = repository;
@@ -32,15 +35,38 @@ public class AuthorController {
     // Create author
     @PostMapping("/authors")
     Author newAuthor(@RequestBody Author newAuthor){
+        newAuthor.validateRequiredAttributes();
         return repository.save(newAuthor);
+    }
+
+    // Update author
+    @PatchMapping("/authors/{id}")
+    Author patchAuthor(@RequestBody Author newAuthor, @PathVariable Long id) {
+        // Check if id from request body is different from the id in the path variable
+        if (newAuthor.getId() != null && !newAuthor.getId().equals(id)){
+            throw new UniqueIdentifierModificationException(newAuthor.getId());
+        }
+        Author existingAuthor = repository.findById(id).orElseThrow(()-> new AuthorNotFoundException(id));
+        if (newAuthor.getName() != null){
+            existingAuthor.setName(newAuthor.getName());
+        }
+        if (newAuthor.getDateOfBirth() != null){
+            existingAuthor.setDateOfBirth(newAuthor.getDateOfBirth());
+        }
+        return repository.save(existingAuthor);
     }
 
     // Replace author
     @PutMapping("/authors/{id}")
-    Author replaceAuthor(@RequestBody Author newAuthor, @PathVariable Long id){
+    Author replaceAuthor(@RequestBody Author newAuthor, @PathVariable Long id) {
+        // Check if id from request body is different from the id in the path variable
+        if (newAuthor.getId() != null && !newAuthor.getId().equals(id)){
+            throw new UniqueIdentifierModificationException(newAuthor.getId());
+        }
+        newAuthor.validateRequiredAttributes();
         return repository.findById(id).map(author -> {
-            author.setName(author.getName());
-            author.setDateOfBirth(author.getDateOfBirth());
+            author.setName(newAuthor.getName());
+            author.setDateOfBirth(newAuthor.getDateOfBirth());
             return repository.save(author);
         }).orElseGet(() -> {
             return repository.save(newAuthor);
@@ -49,7 +75,14 @@ public class AuthorController {
 
     // Delete author
     @DeleteMapping("/authors/{id}")
-    void deleteAuthor(@PathVariable Long id) {
-        repository.deleteById(id);
+    public ResponseEntity<String> deleteAuthor(@PathVariable Long id) {
+        Optional<Author> existingAuthor = repository.findById(id);
+        // Check if author exists before deleting it
+        if(existingAuthor.isPresent()){
+            repository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        // Send 404 if author wasn't found & deleted
+        return ResponseEntity.notFound().build();
     }
 }
